backend/requirements.txt
Flask
mysql-connector-python
pyrebase
pymysql
python-dotenv